first correct implementation

+++ bin/entry_point.sh
@@ -1,18 +1,22 @@
 #!/bin/bash
 
-CONFIG_FILE=_config.yml
+CONFIG_FILE=_config.yml 
 
-/bin/bash -c "rm -f Gemfile.lock && exec jekyll serve --watch --port=8080 --host=0.0.0.0 --livereload --verbose --trace" 
+/bin/bash -c "rm -f Gemfile.lock && exec jekyll serve --watch --port=8080 --host=0.0.0.0 --livereload --verbose --trace"&
 
-inotifywait -m -e modify,move,create,delete $CONFIG_FILE | while read path action file; do
+while true; do
 
-  echo "Change detected to $CONFIG_FILE, restarting Jekyll service"
-  
-  # Kill old Jekyll process
-  jekyll_pid=$(pgrep -f jekyll)
-  kill -KILL $jekyll_pid
+  inotifywait -q -e modify,move,create,delete $CONFIG_FILE
 
-  # Start new Jekyll process
-  /bin/bash -c "rm -f Gemfile.lock && exec jekyll serve --watch --port=8080 --host=0.0.0.0 --livereload --verbose --trace"
+  if [ $? -eq 0 ]; then
+ 
+    echo "Change detecte d to $CONFIG_FILE, restarting Jekyll"
+
+    jekyll_pid=$(pgrep -f jekyll)
+    kill -KILL $jekyll_pid
+
+    /bin/bash -c "rm -f Gemfile.lock && exec jekyll serve --watch --port=8080 --host=0.0.0.0 --livereload --verbose --trace"&
+
+  fi
 
 done
